Race, at least among Racine Students, usually means Black, Latin@, or White.

While folks in town are increasingly likely to identify as multiracial, there are still a small enough number of them that the only statistically useful thing to do is to lump everyone else into "All Other Races." This is obviously bad, but then again, so is all racial categorization? To add to the perfidy, I'm using the "one drop" system that is so perniciously part of American racism: you aren't Latin@ if you are Black, and you're not White if you are Black or Latin@. I am NOT endorsing this system. It reflects the nature of the community fragmentation that is in Racine today.

RACE_LABELS_RACINE

Format

An ordered list of race labels based on recent ancestry from ...

Black

Sub-Saharan Africa

Latin@

Latin America

White

Europe or the Middle East

All Other Races

None of the above