hiRx: Utilities for Higher Expectations' data work

Right now, there are functions for branding, graphical standards, and data wrangling

Branding

Higher Expectations produces reports that use professionally-chosen fonts and color schemes. These include ...

Graphical Standards

Higher Expectations' graphs have a consistent look-and-feel. The following functions facilitate making those graphs.

Wrangling Data

Higher Expectations makes annual summaries of data that come in reasonably consistent formats. Data-wrangling functions capture past efforts to deal with those formats. They are a good place to start when wrangling data from new sources

See also

Author

Maintainer: Ben Taft ben.taft@higherexpectations